Common header for RandomForestModel
